consolidate to one file (datadog.py)
